requirements.txt
pymupdf>=1.23.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
streamlit>=1.34.0
flask>=3.0.0
